pyproject.toml
[tool.poetry]
name = "roombot"
version = "0.1.0"
description = "Roombot!"
authors = ["Zenith <inbox@zenith.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
aurflux = "*"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", branch = "sharding-rework"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
